refactor GLM-Image benchmark, move attention backends benchmark to `kernels/`
