+++ benchmarks/diffusion/recipes/qwen_image_serving_performance.md
