+++ benchmarks/diffusion/recipes/wan_2_2_serving_performance.md
